Update Spritesheet in Windows when Uploading/Writing
When uploading a spritesheet or writing one to the ROM, the window will
update with the new sheet. This makes the modding process a lot easier,
as you know exactly what sheet is currently in the ROM. There isn't an
indication that you haven't saved it yet, but I'll add that later.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>