Optimizations for Palette & Sprite Managers
Most of this code was written years ago and put through a Java to
Kotlin converter, so you know it was horrible.

The Palette Manager will now no longer use those horrible, disgusting
regex splitters and instead uses some super simple string
interpolation.

Reading the values from the ROM in the Sprite Manager used string
splitting, which was awful on performance. Now it uses bit operations
which is a lot faster.

There are still some optimizations that can be made in reading/writing
the ROM, as doing it in batches is faster than byte-by-byte.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>